Make walking skeleton pass.
Flashvnc can receive and send RFB messages from and to the server and has
an architecture that enables it to receive and send events from and to
its GTK GUI.
The socket thread is the main thread, errors and re-connections are handled
here, messages to the user must go through the GUI thread.
Declaratively defined protocol parsers handle most of the syntactic
intricacies of RFB in the setup code and the outer main loop. They are,
however, too weak to help decode the pixel encodings.

Several hacks:
  * Heartbeat messages from server_test_program are a hack to enable a
    timeout because Rust can't do general non-blocking I/O
  * Protocol parsers were originally thought to be very different with much
    less boilerplate code in their definitions. A syntactic
    simplification could be done with closures but you cannot define an
    interface with generic methods using a closure because it cannot
    take generic arguments.
  * Future Rust features like generic array sizes could help very much,
    for example with writing out multi-byte numbers.
  * The write method of parsers takes a value because taking references
    is not possible. One would need something like a "deep reference".