Implement framebuffer.
End-to-end tests need to be run with --test-threads=1 until a solution
is found (running the client in a Xephyr instance maybe).

Implement decoding of raw rectangles into a framebuffer displayed as a
GDK pixbuf.

A benchmark mode measures actual fps and fps as they would be when the
client did not have to wait for the server.

Refactorings:
  * Put GTK-specific stuff into presentation::gtk.
  * Make a method out of the main protocol handler function.
  * Share runner code for the end to end test with the benchmark
    program. As the latter is a cargo bin and the former a
    test, common code must inconveniently be placed in
    tests/benchmark/common.
  * Wrap errors in top-level functions inside a MainError -> errors from
    manual I/O operations can now be easily converted.

Changes:
  * Use xdotool instead of AT-SPI for simulating mouse actions. AT-SPI
    stopped working after an update.