Parallelize decoding.
On a four-core machine, this results in an expected speed-up of about 4
(from 5ms to 1ms-2ms for each frame of the benchmark animation).

Give each worker a jpeg decoder and a zlib decoder. Call the worker
with the correct zlib decoder for zlib data.

Workers write to a FbSlice which is constructed (in an unsafe block)
from a pointer and the FbSize. Locking is done by requiring an FbLock
to be passed to each accept call and ownership to be passed to the
finish method.

further changes:
 * Use AsRawFd instead of IntoRawFd for the setsockopt call.
 * Move resize_fb calls to after decoding to avoid deadlocks.
 * Move Fill decoding to the DecodingMaster.