Completely uncouple decoding.
Move the GUI framebuffer update into another thread pool worker that
first waits for the decoding Future and then copies the framebuffer.
It is joined again at the beginning of the next fb update.

This is thought to give more time to the decoders while the response
fence and the next fb update are exchanged. In practice, it has no
effect, but we keep the change.

Note that in this architecture, it is unwise to let the GUI lock the fb
since the only time the write lock on the fb can be removed is at the
beginning of an fb update and there can be an arbitrary amount of time
between those.

Also a protocol implementation:
 * Cursor pseudo-encoding

refactorings:
 * Move code from the PaletteFilter decoding to a BitBuffer object
   that iterates over single bits in a byte stream.

changes:
 * Drop init function after using it in a ThreadPool.
 * Remove jpeg benchmark measurements. They would have to be moved
   to the DecodingMaster.