Make faster with SDL and consistent pixel formats.
Implement an SDL front-end, which can draw frames with a much more
predictable delay. Note that getting RFB-compatible keysyms is very
hard with SDL.

Change pixel format to BGRX, to match with it with the server and the
GUI. Convert (for GTK), or just copy (for SDL) the FB into a new buffer
for the GUI in a parallelized manner in the thread pool. Use direct
pointer access for that since it was proven to be faster.

Implement a zero-copy mode in which Views with BGRX pixel format (SDL)
simply get the old FB while a new one is allocated. The following FB
update is expected to fill the new buffer completely, which is why the
mode is only turned on after 60 successive full FB updates.

changes:
 * Use sync_channel(5)s for the communication to the front-ends to avoid
   using up excessive amounts of memory when the GUI is not fast enough.
 * Fix scrolling in GTK again (there was a regression).
 * Make the JpegDecoder output BGRX directly.
 * Read command line options (beginning with --). Handle --sdl to
   launch the SDL front-end, otherwise GTK.

refactorings:
 * Encapsulate menu state (f8_pressed, ...) and actions based upon it
   in a Menu object. Used in SDL and GTK.
 * Stop passing FB pointers around. Instead, a ModeLock allows multiple
   writers to the FB as long as they use the same mode (e.g. Decoding).
   This way, the FB can simply be passed to all clients as a
   reference-counted object.
 * Refactor out read_rectangles_and_start_decoding in protocol thread.